requirements.txt
langchain
langchain-community
langchain-groq
langchain-core
langchain-pinecone
langchain-text-splitters
langchain-huggingface
sentence-transformers
pinecone
pypdf
python-dotenv
fastapi
uvicorn